Long -> Double

+++ src/main/java/telraam/station/Fetcher.java
@@ -149,7 +149,7 @@ public class Fetcher {
                             detection.battery,
                             detection.uptimeMs,
                             detection.id,
-                            new Timestamp(detection.detectionTimestamp * 1000)
+                            new Timestamp((long) (detection.detectionTimestamp * 1000))
                     ));
                 }
             }
